Add new feature that allows registering Actors externally
For some types, especially engine built-in classes, registering ActorIO system events and functions is difficult and cumbersome, and most of them need to inherit a separate subclass and implement the IO interface to do so. This new approach allows users to create custom registration classes on the blueprint side and register events and functions to any Actor type.

And for ease of use on the blueprint side, I chose UObject as the registration class instead of Instanced Struct.

The registered class can specify a target class, and when registering subclasses of that class, it will additionally get registered events and functions from the registered class, similar to GetNativeEventsForObject().

Engine internal classes that were originally registered using hard-coded registration can also be refactored using this new registration method, making them easier to maintain in the long run.

For now, I've kept the original registration method and added the new method as a new feature, if the feature is adopted, a larger refactoring can be considered.